% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_agendas.R
\docType{data}
\name{test_agendas}
\alias{test_agendas}
\title{Randomly selected 1,999 Danish local agenda items}
\format{A data frame with 1999 rows and 3 variables:
\describe{
  \item{id}{Unique id for observation}
  \item{text}{Text of the agenda item}
}}
\source{
\url{http://ps.au.dk/forskning/forskningsprojekter/capcas/}
}
\usage{
test_agendas
}
\description{
A dataset containing the text of 1,999 randomly selected agenda items collected and coded by the 
Causes and Policy Consequences of Agenda Setting project
}
\keyword{datasets}

