% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/training_agendas.R
\docType{data}
\name{training_agendas}
\alias{training_agendas}
\title{Randomly selected 2,000 Danish local agenda items}
\format{A data frame with 2000 rows and 3 variables:
\describe{
  \item{id}{Unique id for observation}
  \item{coding}{Policy agendas subtopic coding}
  \item{text}{Text of the agenda item}
}}
\source{
\url{http://ps.au.dk/forskning/forskningsprojekter/capcas/}
}
\usage{
training_agendas
}
\description{
A dataset containing the text and coding of 2,000 randomly selected agenda items collected and coded by the 
Causes and Policy Consequences of Agenda Setting project
}
\keyword{datasets}

