% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topWords.R
\name{topWords}
\alias{topWords}
\title{Find top words by category}
\usage{
topWords(est, latexfile = FALSE, filename = "topwords.tex")
}
\arguments{
\item{est}{The output object from \code{trainNB}}

\item{latexfile}{Logical indicating whether the user wants a latex table of results output into the current working directory.}

\item{filename}{String name for the output file. Defaults to \code{topwords.tex}}
}
\value{
A dataframe with one row for each class. Columns correspond to class name and the top five words associated with the respective class.
}
\description{
Identifies training words (features) most associated with each class in the training data.
}
\examples{
## Load data and create document-feature matrices
train_corpus <- quanteda::corpus(x=training_agendas$text)
train_matrix <- quanteda::dfm(train_corpus,
                    language="danish",
                    stem=TRUE,
                    removeNumbers=FALSE)

est <- trainNB(training_agendas$coding, train_matrix)
words <- topWords(est)

}
\author{
Matt W. Loftis
}

