% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainNB.R
\name{trainNB}
\alias{trainNB}
\title{Train Naive Bayes}
\usage{
trainNB(coding, train_matrix)
}
\arguments{
\item{coding}{The numeric vector of codings}

\item{train_matrix}{A \code{quanteda} document-feature matrix with the number of rows equal to the length of \code{coding}}
}
\value{
A list with the elements
\item{w_0c}{Constant portion of NB classification probabilities.}
\item{w_jc}{Portion of NB classification probabilities that varies with test document word appearances.}
\item{nc}{Frequency of each category in training documents (named numeric vector)}
\item{theta_c}{Unsmoothed prior class probabilities (named numeric vector)}
}
\description{
Trains multiclass Naive Bayes classifier
}
\examples{
## Load data and create document-feature matrices
train_corpus <- quanteda::corpus(x=training_agendas$text)
train_matrix <- quanteda::dfm(train_corpus,
                    language="danish",
                    stem=TRUE,
                    removeNumbers=FALSE)

test_corpus <- quanteda::corpus(x=test_agendas$text)
test_matrix <- quanteda::dfm(test_corpus,
                   language="danish",
                   stem=TRUE,
                   removeNumbers=FALSE)

## Convert matrix of frequencies to matrix of indicators
train_matrix@x[train_matrix@x>1] <- 1
test_matrix@x[test_matrix@x>1] <- 1

## Dropping training features not in the test set
train_matrix <- train_matrix[,(colnames(train_matrix) \%in\% colnames(test_matrix))]

est <- trainNB(training_agendas$coding, train_matrix)

}
\author{
Matt W. Loftis
}

